Changes to work towards ShoopDaLoop standalone deployment with a single entry point.
